+++ src/main/java/com/example/service/CategoryFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Category;
 
 public interface CategoryFetchService {
-	Category fetchById(int id);
+	Category fetchById(Long id);
 
 }
