+++ src/main/java/com/example/service/CategoryUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Category;
 
 public interface CategoryUpdateService {
-	Category updateData(Category category , int id);
+	Category updateData(Category category , Long id);
 
 }
