+++ src/main/java/com/example/service/OwnerFetchService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Owner;
 
 public interface OwnerFetchService {
-	Owner fetchById(int id);
+	Owner fetchById(Long id);
 
 }
