+++ src/main/java/com/example/service/OwnerUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Owner;
 
 public interface OwnerUpdateService {
-	Owner updateData(Owner owner , int id);
+	Owner updateData(Owner owner , Long id);
 
 }
